+++ models/motorcycle.rb
@@ -1,0 +1,9 @@
+require_relative 'vehicle'
+
+class Motorcycle < Vehicle
+
+  def self.num_wheels
+    2
+  end
+  
+end
