+++ spec/spec_helper.rb
@@ -2,3 +2,4 @@ require "rspec"
 require 'bundler/setup'
 require_relative  '../db/setup'
 require_relative "../models/automobile"
+require_relative "../models/motorcycle"
